net_backend: reduce size of TxSegment (#2265)
Shrink `TxSegment` from 48 bytes to 32 bytes by combining flags fields
and reducing the sizes of some fields. This should improve codegen a bit
in the netvsp hot path.

With bigger changes, we could go further here and potentially get down
to 24 or even 16 bytes, e.g., by splitting the head segment to be
metadata-only and putting all payload info on the tail segments. Don't
do this just yet.

+++ vm/devices/net/net_dio/src/lib.rs
@@ -178,7 +178,7 @@ impl Queue for DioQueue {
             while !segments.is_empty() {
                 let (metadata, this, rest) = next_packet(segments);
                 segments = rest;
-                nic.write_with(metadata.len, |mut buf| -> anyhow::Result<_> {
+                nic.write_with(metadata.len as usize, |mut buf| -> anyhow::Result<_> {
                     for segment in this {
                         let (this, rest) = buf.split_at_mut(segment.len as usize);
                         mem.read_at(segment.gpa, this)
